backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
google-cloud-firestore
google-generativeai
openai
pydantic
pydantic-settings
email-validator
requests
python-multipart
passlib[bcrypt]
firebase-admin

